requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
plotly==4.14.3
requests==2.25.1
urllib3==1.26.4
wincertstore==0.2
python-dotenv==0.17.0
abstract_python_core==0.1.0
